+++ src/app/core/directives/authorization-toggle.directive.ts
@@ -4,6 +4,19 @@ import { distinctUntilChanged, takeUntil } from 'rxjs/operators';
 
 import { AuthorizationToggleService } from 'ish-core/utils/authorization-toggle/authorization-toggle.service';
 
+/**
+ * Structural directive.
+ * Used on an element, this element will only be rendered if the logged in user has the specified permission or one of the specified permission in case an array is given.
+ *
+ * @example
+ * <div *ishIsAuthorizedTo="APP_B2B_PURCHASE'">
+ *   Only visible when the current user is allowed to purchase items.
+ * </div>
+ * or
+ * <div *ishIsAuthorizedTo="['APP_B2B_ORDER_APPROVAL', 'APP_B2B_MANAGE_COSTCENTER']">
+ *   Only visible when the current user has at least one of the given permissions.
+ * </div>
+ */
 @Directive({
   selector: '[ishIsAuthorizedTo]',
 })
@@ -29,7 +42,7 @@ export class AuthorizationToggleDirective implements OnDestroy {
     });
   }
 
-  @Input() set ishIsAuthorizedTo(permission: string) {
+  @Input() set ishIsAuthorizedTo(permission: string | string[]) {
     // end previous subscription and subscribe to new permission
     if (this.subscription) {
       // tslint:disable-next-line: ban